feat: add reaction to spaces in example app (#58)
* feat: add reaction to spaces in example app

* refactor: replace positional params with named params

* chore: fix typo in function name

---------
