+++ example/lib/views/spaces/__spaces.dart
@@ -1,4 +1,5 @@
 export 'create_space_screen.dart';
-export 'live_space_screen.dart';
+export 'live_space/live_space_screen.dart';
 export 'trending_space_screen.dart';
 export 'my_spaces_screen.dart';
+export 'live_space/live_space_provider.dart';
