+++ lib/src/user/src/get_user.dart
@@ -14,7 +14,7 @@ Future<User?> getUser({
     return null;
   }
 
-  if(result is String){
+  if (result is String) {
     throw Exception("Unable to get user");
   }
