+++ lib/src/values/src/constants.dart
@@ -97,7 +97,8 @@ class CHAT {
   static const String META_SPACE_START = 'START_SPACE';
   static const String META_SPACE_LISTENER_ADD = 'ADD_LISTENER';
   static const String META_SPACE_LISTENER_REMOVE = 'REMOVE_LISTENER';
-  static const String META_SPACE_LISTENER_PRIVILEGE = 'ASSIGN_LISTENER_PRIVILEGE';
+  static const String META_SPACE_LISTENER_PRIVILEGE =
+      'ASSIGN_LISTENER_PRIVILEGE';
   static const String META_SPACE_SPEAKER_PRIVILEGE = 'ASSIGN_SPEAKER_PRIVILEGE';
   static const String META_SPACE_COHOST_PRIVILEGE = 'ASSIGN_COHOST_PRIVILEGE';
   static const String REACTION_THUMBSUP = '\u{1F44D}';
